refactor #187 : ProductStatus를 vo 패키지로 이동
