+++ product-service/src/main/java/com/musinsam/productservice/application/dto/request/ReqProductPostDtoApiV1.java
@@ -1,7 +1,7 @@
 package com.musinsam.productservice.application.dto.request;
 
 import com.musinsam.productservice.domain.product.entity.ProductEntity;
-import com.musinsam.productservice.domain.product.entity.ProductStatus;
+import com.musinsam.productservice.domain.product.vo.ProductStatus;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
