+++ product-service/src/main/java/com/musinsam/productservice/application/service/ProductServiceImplApiV1.java
@@ -12,9 +12,9 @@ import com.musinsam.productservice.application.integration.CouponClientApiV1;
 import com.musinsam.productservice.application.integration.ShopClientApiV1;
 import com.musinsam.productservice.domain.product.entity.ProductEntity;
 import com.musinsam.productservice.domain.product.entity.ProductImageEntity;
-import com.musinsam.productservice.domain.product.entity.ProductStatus;
 import com.musinsam.productservice.domain.product.repository.ProductImageRepository;
 import com.musinsam.productservice.domain.product.repository.ProductRepository;
+import com.musinsam.productservice.domain.product.vo.ProductStatus;
 import com.musinsam.productservice.infrastructure.dto.res.ResShopCouponDtoApiV1;
 import com.musinsam.productservice.infrastructure.s3.S3Folder;
 import com.musinsam.productservice.infrastructure.s3.service.S3Service;
